fix(qa-sprint): 16 bugs críticos antes da apresentação a investidores
P0 — detalhe/edição de produto 100% quebrado:
- VariacaoDetalhe.cs e ProdutoDetalhe.cs: int → decimal em campos de
  quantidade (numeric(18,3) serializa como 5.000, falha em int)

P0 — rotas inexistentes retornando 404/UUID error:
- PedidosController: adiciona GET /pedidos/novo → redirect para Index
- ClientesController: adiciona GET /clientes/novo → redirect para Index

P0 — histórico de entradas/saídas explodindo com 500:
- MovimentacaoEstoqueRepository: DateTime.SpecifyKind(Utc) nos filtros
  de data (Npgsql rejeita Kind=Unspecified em colunas timestamptz)

P0 — /dispositivos, /lotes-mobile e /clientes-mobile retornando 500:
- MobileDevicesController: await sequencial + try-catch isolado no
  health call (pode lançar sincronicamente antes de retornar Task)
- LotesMobileController e ClientesMobileController: try-catch defensivo

P1 — gráfico receita com apenas 1 ponto (mês atual):
- ReceitaAnalyticsQueries: DateTime.SpecifyKind para GetReceitaPorPeriodo
  e GetVendasPorCanal; também corrige DashboardAnalyticsQueries

P2 — UX/visual:
- Dashboard/Index.cshtml: hidden lg:block nas Ações Rápidas (duplicadas
  com bottom nav em mobile); guard GraficoDados.Any(d>0) no chart
- Analytics/Index.cshtml: mesmo guard no gráfico de receita
- Saidas/Nova.cshtml: valorUnit = precoReferencia || custoReferencia
  (precoVendaSugerido e custoUnitario não existem no response)
- Lotes/Detail.cshtml: tab Etiquetas usa l.TotalUnidades
- Clientes/Detail.cshtml: filtra registros de sync interno do log

+++ EasyStock.Web/Models/Api/ProdutoDetalhe.cs
@@ -22,7 +22,7 @@ public record ProdutoDetalhe
     public decimal? PrecoReferencia { get; init; }
     public decimal? MargemEstimada { get; init; }
     public ProdutoDimensoesDetalhe? Dimensoes { get; init; }
-    public int QuantidadeTotalEstoque { get; init; }
+    public decimal QuantidadeTotalEstoque { get; init; }
     public DateTime? UltimaEntradaEm { get; init; }
     public List<ProdutoFotoDetalhe> Fotos { get; init; } = [];
     public List<VariacaoDetalhe> Variacoes { get; init; } = [];
